Reorganize API module layout
