+++ src/api/agent/userProfile.js
@@ -1,5 +1,5 @@
 /* global chrome */
-import { textComplete, getLLMConfigForMemory } from "./llmTextComplete";
+import { textComplete, getLLMConfigForMemory } from "../llm/providers/textComplete";
 
 // ==================== Storage keys ====================
