+++ src/api/llm/core/reasoning.js
