+++ src/api/llm/providers/anthropic.js
@@ -1,10 +1,10 @@
-import { resolveLlmRequestUrl } from "../llmEndpoint";
-import { DEFAULT_ANTHROPIC_CACHE_CONTROL } from "./constants";
-import { API_TYPES } from "./config";
-import { buildFirstPacketTimeoutError, createFirstPacketTimeoutState, createLlmStreamError, getFirstPacketTimeoutMs, isAbortError, mergeUsage } from "./shared";
-import { getTools } from "./tools";
-import { isLongToolArgumentName } from "./longToolArgs";
-import { buildAnthropicReasoningFields } from "./reasoning";
+import { resolveLlmRequestUrl } from "../core/endpoint";
+import { DEFAULT_ANTHROPIC_CACHE_CONTROL } from "../core/constants";
+import { API_TYPES } from "../core/config";
+import { buildFirstPacketTimeoutError, createFirstPacketTimeoutState, createLlmStreamError, getFirstPacketTimeoutMs, isAbortError, mergeUsage } from "../core/shared";
+import { getTools } from "../tools/definitions";
+import { isLongToolArgumentName } from "../core/longToolArgs";
+import { buildAnthropicReasoningFields } from "../core/reasoning";
 
 // ==================== Anthropic Messages API ====================
 
@@ -308,4 +308,3 @@ function extractAnthropicStreamUsage(event) {
 function mergeAnthropicUsage(current = {}, next) {
   return mergeUsage(current, next);
 }
-
