+++ src/api/llm/providers/openai-responses.test.js
