+++ src/api/llm/tools/builtins/_shared.js
