+++ src/api/llm/tools/builtins/downloadHelper.js
