+++ src/api/llm/tools/builtins/groups.js
