+++ src/api/llm/tools/builtins/history.js
