+++ src/api/llm/tools/builtins/imageApi.js
