+++ src/api/llm/tools/builtins/macros.js
