+++ src/api/llm/tools/builtins/misc.js
