+++ src/api/llm/tools/builtins/tabs.js
