+++ src/api/llm/tools/builtins/windows.js
